// src/lib/types.ts
export type CompanyStatus = 'Producer' | 'Developer' | 'Explorer' | 'Royalty' | 'Other';
// ...
export interface SortState {
  key: string;
  direction: 'asc' | 'desc';
}